requirements.txt
asttokens==2.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
filelock==3.8.0
fonttools==4.38.0
gdown==4.5.3
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
jupyter_client==7.4.5
jupyter_core==5.0.0
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
monai-weekly==1.1.dev2245
nest-asyncio==1.5.6
nibabel==4.0.2
numpy==1.23.4
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.3
prompt-toolkit==3.0.32
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-ignite==0.4.10
pywin32==305
pyzmq==24.0.1
requests==2.28.1
scipy==1.9.3
SimpleITK==2.2.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.0
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
